Fix the initializer for Rails < 7.2
- Now that Campaigns developers will have their own users to connect to
Snowflake, the initializer needs to pass the config instead of
connection_options so we can filter tables and views by database and
schema.
- With the fix above, we can fetch only the tables the user has access
on the database and schema the connection was configured, this fix the
problem when `show tables in account` was called for privileged users
as they have access to way more than 10000 tables, which causes
Snowflake to return an error due to the size of the recordset being
returned.
- Update the code to match the github/BookBub/odbc_adapter which was
used as a base for the Rails 7.2 fixes.