add `allow_retry` optional arg for rails 7.2 compatibility
Rails 7.1 introduced the ability to specify an `allow_retry` option,
allowing queries to be retried per the databases configured retries
value if a connection-related error occurs [1]. This option appears
to be explicitly passed over through to `internal_exec_query` as of
Rails 7.2, resulting in argument errors.

This updates `internal_exec_query` to accept an optional `allow_retry`
argument so that it is compatible with Rails 7.2+ where such an argument
may be forwarded to the method.

[1] https://github.com/rails/rails/blob/40a0294faf6c8aefdbad16d88d3b94fa3050f6dd/activerecord/CHANGELOG.md?plain=1#L1844-L1845