explicitly define quote_table_name for adapter (rails 7.2 compatibility)
The abstract adapter was reworked in Rails 7.2+ to define its table and
column quoting logic at the class level, eliminating the need for an
active connection instance to access said quoting logic - this change
was introduced via https://github.com/rails/rails/pull/51174

As a consequence of the above mentioned rework, the `quote_table_name`
and `quote_table_column` methods aren't accessible in the same way to
subclasses in Rails 7.2+ as they were in prior versions. This results
in table names being quoted inconsistent with column names, as the
latter is explicitly defined on the odbc adapter while the former
is not.

This explicitly defines the `quote_table_name` method for the ODBC
adapter to ensure consistent behavior between Rails versions (i.e
7.1, 7.2, and beyond).

Note: This approach was taken as the most minimally intrusive change
to the adapter, but if at some point the gem drops support for Rails
7.1, this Quoting class can be tweaked to be more in line with the
newer approach taken for adapters in Rails 7.2+, as shown in the
above referenced PR. Specifically, the `quote_column_name` can
be defined as a class method, allowing the superclass definitions
to take care of the rest, e.g.:

```ruby
module ODBCAdapter
  module Quoting
    extend ActiveSupport::Concern

    module ClassMethods
      def quote_column_name(name)
        # ...
      end
    end

    # ...
  end
end
```

Though looking at the current `quote_column_name` implementation for
this adapter, it might need to be reworked to not depend on an instance-
level connection, as it currently appears to require one for accessing
various database metadata info, see:

- https://github.com/doximity/odbc_adapter/blob/724e76261e8eb0173cd80477696a66c54f8b1e97/lib/odbc_adapter/quoting.rb#L11
- https://github.com/doximity/odbc_adapter/blob/724e76261e8eb0173cd80477696a66c54f8b1e97/lib/odbc_adapter/quoting.rb#L20

For a comparison reference beween pre/post Rails 7.2 quoting implementations:

- Rails 7.1: https://github.com/rails/rails/blob/7-1-stable/activerecord/lib/active_record/connection_adapters/abstract/quoting.rb
- Rails 7.2: https://github.com/rails/rails/blob/7-2-stable/activerecord/lib/active_record/connection_adapters/abstract/quoting.rb