call preprocess_query if defined, fall back to transform_query (rails 8 support)
As of Rails 8.0, Active Record's adapter internals have been refactored
to streamline some of their interfaces for better consistency [1]. This
refactor removed the private `#transform_query` method in favor of a
new `#preprocess_query` method that handles a few extra checks in
addition to running transformers.

This change checks for the presence of a `preprocess_query`  method
and runs it if defined, otherwise falling back to the pre Rails 8
`transform_query` method.

[1] https://github.com/rails/rails/pull/52428