define write_query? (rails 8.1 compatibility)
As of Rails 8.1, some of Active Record's adapter internals have been
refactored to eliminate duplicate `write_query?` method calls via [1].
As a result, `write_query?` now actually gets called in this adapter
as it runs immediately when `preprocess_query` is called and then
raises via `NotImplementedError` as it was not being defined -
this wasn't an issue before as previously the check seems to
have been short circuiting at `preventing_writes?` [2].

This defines `write_query?` for this adapter, using the same approach
taken in the officially supported adapters (e.g. postgresql [3]).

The args passed to the read query regex build method are for commands
that Snowflake supports as per their documentation [4][5][6][7][8].

[1] https://github.com/rails/rails/commit/07854a55607f845a72c4a9d43d4bedfe954d038c
[2] https://github.com/rails/rails/blob/529f933fc8b13114d308dd0752f76a9e293c8537/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb#L196
[3] https://github.com/rails/rails/blob/529f933fc8b13114d308dd0752f76a9e293c8537/activerecord/lib/active_record/connection_adapters/postgresql/database_statements.rb

[4] https://docs.snowflake.com/en/sql-reference/snowflake-scripting/close
[5] https://docs.snowflake.com/en/sql-reference/snowflake-scripting/declare
[6] https://docs.snowflake.com/en/sql-reference/snowflake-scripting/fetch
[7] https://docs.snowflake.com/en/sql-reference/sql/set
[8] https://docs.snowflake.com/en/sql-reference/sql/show