pass cast type to column (rails 8.1 compatibility)
As of Rails 8.1, the ActiveRecord adapter internals
have been reworked to require that instances of
`ActiveRecord::ConnectionAdapters::Column` receive
a `cast_type` argument for instantiation (introduced
via [1]).

This updates the adapter code to pass the cast type value
to Column as an instantiation argument, which should fix
undefined method errors seen as of Rails 8.1.

Changing the method signature like this means that these
changes would not be backward compatible with Rails versions
lower than 8.1, and per the gemspec this gem is stated to
support Rails 7.1.2+. We can either bump the min required
activerecord dependency to 8.1 or tweak this so that both
pre and post 8.1 signatures are supported.

[1] https://github.com/rails/rails/commit/9ad36e067222478090b36a985090475bb03e398c