support column initialization for both pre and post ActiveRecord 8.1.0
As of Rails 8.1, the ActiveRecord adapter internals
have been reworked to require that instances of
`ActiveRecord::ConnectionAdapters::Column` receive
a `cast_type` argument for instantiation (introduced
via [1]). We made changes to adjust this in a previous
commit but it would mean that apps using ActiveRecord
older than 8.1.0 would not be compatible as-is.

This reworks the Column initialization so that it essentially
forwards the positional arguments over to its superclass and
still allows the assigning of the additional attributes that
the adapter's column class requires. The positional args are
then constructed based on the ActiveRecord version so that
both pre and post Rails 8.1.0 are supported.

[1] https://github.com/rails/rails/commit/9ad36e067222478090b36a985090475bb03e398c